op-e2e: Extract init logic to a separate helper file.
Remove duplication for verboseGethNodes support
